+++ examples/data-sources/clearpass_role_mapping/data-source.tf
@@ -1,0 +1,7 @@
+data "clearpass_role_mapping" "by_id" {
+  id = 12
+}
+
+data "clearpass_role_mapping" "by_name" {
+  name = "Guest User Mapping"
+}